Move keyboard actions into their own file
