+++ client-v2/src/actions/ArticleFragments.ts
@@ -8,16 +8,11 @@ import {
   articleFragmentsReceived,
   maybeAddFrontPublicationDate
 } from 'shared/actions/ArticleFragments';
-import {
-  nextIndexAndGroupSelector,
-  nextClipboardIndexSelector
-} from '../selectors/keyboardNavigationSelectors';
 import { ArticleFragment } from 'shared/types/Collection';
 import {
   selectSharedState,
   articleFragmentsSelector,
-  groupSiblingsArticleCountSelector,
-  indexInGroupSelector
+  groupSiblingsArticleCountSelector
 } from 'shared/selectors/shared';
 import { ThunkResult, Dispatch } from 'types/Store';
 import { addPersistMetaToAction } from 'util/storeMiddleware';
@@ -254,52 +249,6 @@ const removeArticleFragment = (
   };
 };
 
-const keyboardArticleFragmentMove = (
-  fragment: ArticleFragment,
-  action: 'up' | 'down',
-  persistTo: 'collection' | 'clipboard',
-  groupId?: string
-): ThunkResult<void> => {
-  return (dispatch: Dispatch, getState) => {
-    const state = getState();
-    const id = fragment.uuid;
-    if (persistTo === 'collection') {
-      const fromIndex = indexInGroupSelector(
-        selectSharedState(state),
-        groupId || '',
-        id
-      );
-      const type = 'group';
-
-      const from: PosSpec = { type, index: fromIndex, id: groupId || '' };
-
-      const nextPosition = nextIndexAndGroupSelector(
-        state,
-        id,
-        groupId || '',
-        action
-      );
-
-      if (nextPosition && nextPosition.nextGroupId) {
-        const { toIndex, nextGroupId } = nextPosition;
-
-        const to: PosSpec = { type, index: toIndex, id: nextGroupId };
-
-        dispatch(moveArticleFragment(to, fragment, from, persistTo));
-      }
-    } else if (persistTo === 'clipboard') {
-      const clipboardIndeces = nextClipboardIndexSelector(state, id, action);
-      if (clipboardIndeces) {
-        const { fromIndex, toIndex } = clipboardIndeces;
-        const type = 'clipboard';
-        const from = { type, index: fromIndex, id: 'clipboard' };
-        const to = { type, index: toIndex, id: 'clipboard' };
-        dispatch(moveArticleFragment(to, fragment, from, persistTo));
-      }
-    }
-  };
-};
-
 const moveArticleFragment = (
   to: PosSpec,
   fragment: ArticleFragment,
@@ -359,6 +308,5 @@ export {
   updateArticleFragmentMetaWithPersist as updateArticleFragmentMeta,
   updateClipboardArticleFragmentMetaWithPersist as updateClipboardArticleFragmentMeta,
   removeArticleFragment,
-  addImageToArticleFragment,
-  keyboardArticleFragmentMove
+  addImageToArticleFragment
 };
